:recycle: 매칭 후 문자 전송

+++ src/main/java/animal/meeting/domain/user/entity/SecretKey.java
@@ -6,20 +6,20 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class SecretKey {
-	private final long secretKey1;
-	private final long secretKey2;
-	private final long secretKey3;
+	private final int secretKey1;
+	private final int secretKey2;
+	private final int secretKey3;
 
 	public SecretKey(
-		@Value("${SECRET_KEY_1}") long secretKey1,
-		@Value("${SECRET_KEY_2}") long secretKey2,
-		@Value("${SECRET_KEY_3}") long secretKey3) {
+		@Value("${SECRET_KEY_1}") int secretKey1,
+		@Value("${SECRET_KEY_2}") int secretKey2,
+		@Value("${SECRET_KEY_3}") int secretKey3) {
 		this.secretKey1 = secretKey1;
 		this.secretKey2 = secretKey2;
 		this.secretKey3 = secretKey3;
 	}
 
-	public boolean isValidSecretKey(long inputKey) {
+	public boolean isValidSecretKey(int inputKey) {
 		return inputKey == secretKey1 || inputKey == secretKey2 || inputKey == secretKey3;
 	}
 }
